Refactor: 투두리스트 디렉토리 구조 개선 및 성능 최적화 (#225)
* #221 refactor: 투두 관련 스타일 분리

* fix: 투두 삭제 mutate 함수 오타 수정 및 options를 옵셔널로 변경

* #221 feat: 투두 수정 시 쿼리 파라미터를 Context로 전달하도록 변경

* #221 refactor: 재사용 스타일을 common 폴더로 분리

* #221 refactor: 투두 수정 및 삭제 로직을 커스텀 훅으로 분리

* #221 refactor: 투두 디렉토리 구조 개선 및 최적화

- 'Todo' 접두사 제거
- React.memo를 활용하여 리렌더링 최적화
- 투두 상태 스타일을 메모이제이션하여 성능 개선

* #221 refactor: 투두 조회 로직과 헤더를 Container 컴포넌트로 통합

- TodoList와 Header를 하나의 Container 컴포넌트로 병합
- 투두 조회 로직을 별도 모듈로 분리
- Context를 제거하고 필요한 상태를 props로 전달하도록 변경
