+++ src/components/Todo/Form.tsx
@@ -1,15 +1,15 @@
-import { Form } from '@/components/common';
+import { Form as TodoForm } from '@/components/common';
+import { formStyle } from '@/components/Todo/styles';
 import { TODO_TYPES } from '@/constants/todo';
 import { useToastHandler } from '@/hooks';
 import { useCreateTodo } from '@/hooks/mutations';
 import { todoKeys } from '@/hooks/queries';
 import { InfiniteQueryData } from '@/hooks/queries/types';
 import { ApiResponse, PostTodoRequest, ToDoDetail } from '@/types';
-import { css } from '@emotion/react';
 import { useQueryClient } from '@tanstack/react-query';
-import { FormEvent, KeyboardEventHandler, useCallback, useState } from 'react';
+import { FormEvent, KeyboardEventHandler, memo, useState } from 'react';
 
-const TodoForm = ({
+const Form = ({
   squadId,
   selectedDay,
   squadMemberId,
@@ -51,16 +51,16 @@ const TodoForm = ({
     setContents('');
   };
 
-  const handleEnterSubmit: KeyboardEventHandler<HTMLFormElement> = useCallback((e) => {
+  const handleEnterSubmit: KeyboardEventHandler<HTMLFormElement> = (e) => {
     if (e.key !== 'Enter') return;
     if (!e.nativeEvent.isComposing) {
       e.preventDefault();
       handleSubmit(e);
     }
-  }, []);
+  };
 
   return (
-    <Form
+    <TodoForm
       type="text"
       onSubmit={handleSubmit}
       onKeyDown={handleEnterSubmit}
@@ -72,8 +72,4 @@ const TodoForm = ({
   );
 };
 
-export default TodoForm;
-
-const formStyle = () => css`
-  margin: 16px;
-`;
+export default memo(Form);
