+++ src/context/todo/context.ts
@@ -1,0 +1,4 @@
+import { TodoQueryParams } from '@/types';
+import { createContext } from 'react';
+
+export const TodoQueryParamsContext = createContext<TodoQueryParams | null>(null);
